Fixing Linux compilation

+++ lava/Image.cpp
@@ -33,7 +33,7 @@ namespace lava
 
     vk::MemoryRequirements memReqs = static_cast< vk::Device >( *_device ).getImageMemoryRequirements( _image );
     uint32_t memoryTypeIndex = findMemoryType( _device->_physicalDevice->getMemoryProperties( ), memReqs.memoryTypeBits, _memoryPropertyFlags );
-    assert( memoryTypeIndex != -1 );
+    assert( memoryTypeIndex != uint32_t( -1 ) );
     auto mem = _device->allocateMemReqMemory( memReqs, _memoryPropertyFlags );
     vk::Device( *_device ).bindImageMemory( _image, mem, 0 );
   }
